feat: add utility classes for dealing with non-standard PVs
In Mobipix, we have a couple awkward PVs to deal with in Ophyd:

1. Some PVs accept a float, but the RBV is of type string. This makes
the default EpicsSignal behavior not work, as the comparison is never true.
What's more, if you try to set a tolerance for them, numpy errors out,
since it is thrown incompatible types at.

2. Some PVs have an RBV, but function as a Command-Return schema. So,
for instance, there's one signal that you send an empty string to,
and the RBV gets set to "Done" when it finishes the command. Ophyd always
tries to check the RBV for the same value as you set, so it nevers
finishes waiting for it.

To work around those, I created the "EpicsSignalWithCustomReadout" signal,
that extends EpicsSignal with the ability to 1. cast the expected and
current values to a common type, and 2. set a custom value to check in the
read_pv for completion.

The design of _LooseComparator is a bit awkward, but I don't currently have
the knowledge of how to make it better.

Signed-off-by: Sofia Donato Ferreira <sofia.ferreira@lnls.br>