feat(AD): add Mobipix V1 ophyd class
This is not an exaustive class with absolutely all functionality
available in the Mobipix V1 IOC, but it covers the energy threshold scan
application for now.

The 'MobipixEnergyThresholdSetter' class is simply a workaround for
providing a Motor-like interface to it, so we can pass it directly to
a Bluesky plan as a step-scan, or the like.

Signed-off-by: Sofia Donato Ferreira <sofia.ferreira@lnls.br>