Revert "refactor: switch compacteras channels and scales to dynamic ones"
This reverts commit d78d8a46250f4973370420b81749abe23fef750c.

This revert is being done because the 'add_components_to_device' function
appears to be super slow, contributing to a big performance hit on the
CompactERAS interface.

Un-reverting this would require either the aforementioned function to
be optimized, the CompactERAS device to switch to ophyd-async, or the
GUI to depend less on the SoftIOC device instantiation.

Signed-off-by: Sofia Donato Ferreira <sofia.ferreira@lnls.br>