fix: properly implement trigger logic for CRIO devices
Because we have little control over the sampling period of CRIO devices,
its trigger logic is a bit weird. We must wait a full `PvAvgTime` period
before reading the device, because it is doing an acquisition.

Furthermore, we often use CRIO devices with their individual signals, so
we must take care to only wait that sampling period once for scan point,
otherwise too much time would be lost.

This uses a heuristic, in which a "scan point" is detected as not reading
the same signal twice. When a repeated read is made, we assume it is
referent to another scan point, and re-do the global trigger.

Signed-off-by: Sofia Donato Ferreira <sofia.ferreira@lnls.br>