feat: add IVU and HD-DCM devices
The IVU device may suffer from some performance issues due to the double
actuate on the KParamChange-Cmd signal. If that's the case, maybe we should
turn it into a standalone signal.

The HD-DCM device is entirely focused on the follower mode operation in this
implementation, so other configurations are not supported just yet.

Signed-off-by: Sofia Donato Ferreira <sofia.ferreira@lnls.br>