fix: do not override tolerance fields in EpicsSignalWithCustomReadout
If we do override these fields, as was done previously, the expected
readout objects created later by this instance will not take into
account the tolerance.

Also, we must not call _wait_for_value with a tolerance set, as that
overrides a None tolerance passed as argument in that function.

Signed-off-by: Sofia Donato Ferreira <sofia.ferreira@lnls.br>