fix!: replace VirtualControllableMotor for a factory function
With a recent change in Ophyd, component instantiation now passes
through a 'issubclass' check, which cannot be used with functions.

That means that our previous method of creating Components using
functions as classes doesn't work anymore (and was kind of a hack).

This patch changes it so that we use a factory function instead,
which returns you a class type object to use in Components, ensuring
we use Components as they were originally intended (and mainly tested for).

By consequence, this patch also breaks client code using
VirtualControllableMotor's directly.

Signed-off-by: Sofia Donato Ferreira <sofia.ferreira@lnls.br>