fix: manually exclude broken 'units' metadata in descriptor for Pimega
Some Pimega signals, like TriggerMode or ImageMode, do not have EGU values,
and as such the returned 'units' value is None.

There are other signals that have this property and don't cause issues
(see EpicsMotor's user_offset_dir, for instance), so I'm not sure why
we're having problems here. Anyway, this seems to fix the issue for now.

Signed-off-by: Sofia Donato Ferreira <sofia.ferreira@lnls.br>