feat: error out when Pimega fails to start acquisition
This prevents plans from being stuck in trigger, because when an
error occurs, the acquire signal never goes high.

The previous attempt at this, using IOCStatusMessage to detect
a failure state, as mentioned in code review, wasn't optimal in
the slightest, and also possibly wouldn't catch all such errors.

In this patch, we instead use the Acquire signal itself to determine
when an error occurs. This relies on the acquisition signal setpoint
changing value from 'Done' to 'Acquire' when the state is updated on
the IOC, as otherwise we won't get a callback to react to, and we won't
be able to check the alarm state.

This seems to be the only reasonable source of updates we get from the IOC.

Signed-off-by: Sofia Donato Ferreira <sofia.ferreira@lnls.br>