feat: improve error printout of Pimega acquisition exceptions
This adds the argument information we provide to the exception we
created into the end of the traceback, instead of it being left only
at the top of the error message.

This is needed because ADTriggerStatus inherits from DeviceStatus,
which implements a __str__ method that doesn't print the 'args'
property we configured.

The custom exception class is intended to provide a slightly clearer
traceback message to the user.

Signed-off-by: Sofia Donato Ferreira <sofia.ferreira@lnls.br>