feat: add 'install_package' util function to install pip packages
We previously had this function created locally in each sophys-server
repository, each doing its own thing. For maintenance, it seemed better
to have it here, and have the base queueserver image come with this
package installed with this function available already.

Furthermore, compared with that version, this one allows us to disable
the pip cache, which is responsible for some unwanted disk usage on
our container images.

Signed-off-by: Sofia Donato Ferreira <sofia.ferreira@lnls.br>