fix: add common sample rate to base device
Since this "common" sample rate just sets all the other channels
to the same value as common, it would be a bit confusing to have
this component as config/hinted. Every channel in this case is now
resposible for its own configuration, and common_sample_rate is
intented to be used inside plans to set all channels to the same
sample rate.