fix: update consumer offsets before starting producers
It seems like kafka-python(-ng) does not retrieve the current offsets
before actually starting to listen on the topic partition. This causes
the monitor to potentially miss events from a producer when everything
happens too fast, such as in the case of a local test with a local Kafka
broker.

The solution proposed here is to actually update those offsets before
allowing the tests to go on, ensuring the monitor points to the correct
offset, and doesn't skip any events.

Signed-off-by: Sofia Donato Ferreira <sofia.ferreira@lnls.br>