refactor: PicoloAcquisitonTime segregate reponsabilities
As it was implemented, the (override) set method knew how validate
and format acquire time. Now a TimeBase class does the dirty job
and a validator check if the produced string is valid. This will
enable writing tests to this part of the code, without having to
run an instance of the IOC and validate the user inputs