meta: remove numpy dependency restriction
According to the merge request that introduced this change:

"Numpy version above 2.0 breaks the numpydocs package required by the Queueserver"

This is not corroborated by upstream reports, and shouldn't be a tie on
this package's dependencies. Rather, if there is such an issue, it should
be solved by a version tie downstream, or fixed upstream (preferably).

Signed-off-by: Sofia Donato Ferreira <sofia.ferreira@lnls.br>